Move API and tests to partcipants app
