+++ apps/challenges/urls.py
@@ -206,9 +206,4 @@ urlpatterns = [
         views.finish_annotation_file_upload,
         name="finish_annotation_file_upload",
     ),
-    url(
-        r"^challenge/(?P<challenge_pk>[0-9]+)/team/(?P<participant_team_pk>[0-9]+)/remove_participant_team/$",
-        views.remove_participant_team_from_challenge,
-        name="remove_participant_team_from_challenge",
-    ),
 ]
